requirements.txt
flask
azure-identity
azure-storage-blob
azure-ai-projects
reportlab
werkzeug

